remove hash_table module
