--- core/src/hash_table/actor.rs
@@ -1,261 +1,0 @@
-use hash_table::HashTable;
-use holochain_core_types::{
-    actor::{AskSelf, Protocol, SYS},
-    cas::content::Address,
-    entry::Entry,
-    entry_meta::EntryMeta,
-    error::HolochainError,
-    keys::Keys,
-};
-use riker::actors::*;
-use snowflake;
-
-// anything that can be asked of HashTable and block on responses
-// needed to support implementing ask on upstream ActorRef from riker
-pub trait AskHashTable: HashTable {}
-
-impl AskHashTable for ActorRef<Protocol> {}
-
-impl HashTable for ActorRef<Protocol> {
-    fn setup(&mut self) -> Result<(), HolochainError> {
-        let response = self.block_on_ask(Protocol::Setup)?;
-        unwrap_to!(response => Protocol::SetupResult).clone()
-    }
-
-    fn teardown(&mut self) -> Result<(), HolochainError> {
-        let response = self.block_on_ask(Protocol::Teardown)?;
-        unwrap_to!(response => Protocol::TeardownResult).clone()
-    }
-
-    fn put_entry(&mut self, entry: &Entry) -> Result<(), HolochainError> {
-        let response = self.block_on_ask(Protocol::PutEntry(entry.clone()))?;
-        unwrap_to!(response => Protocol::PutEntryResult).clone()
-    }
-
-    // TODO #358 - Rename entry to something else. Entry is only to be used for a chain.
-    // https://github.com/holochain/holochain-rust/pull/340#discussion_r220332130
-    fn entry(&self, address: &Address) -> Result<Option<Entry>, HolochainError> {
-        let response = self.block_on_ask(Protocol::GetEntry(address.clone()))?;
-        unwrap_to!(response => Protocol::GetEntryResult).clone()
-    }
-
-    fn modify_entry(
-        &mut self,
-        keys: &Keys,
-        old_entry: &Entry,
-        new_entry: &Entry,
-    ) -> Result<(), HolochainError> {
-        let response = self.block_on_ask(Protocol::ModifyEntry {
-            keys: keys.clone(),
-            old: old_entry.clone(),
-            new: new_entry.clone(),
-        })?;
-        unwrap_to!(response => Protocol::ModifyEntryResult).clone()
-    }
-
-    fn retract_entry(&mut self, keys: &Keys, entry: &Entry) -> Result<(), HolochainError> {
-        let response = self.block_on_ask(Protocol::RetractEntry {
-            keys: keys.clone(),
-            entry: entry.clone(),
-        })?;
-        unwrap_to!(response => Protocol::RetractEntryResult).clone()
-    }
-
-    fn assert_meta(&mut self, meta: &EntryMeta) -> Result<(), HolochainError> {
-        let response = self.block_on_ask(Protocol::AssertMeta(meta.clone()))?;
-        unwrap_to!(response => Protocol::AssertMetaResult).clone()
-    }
-
-    fn get_meta(&mut self, address: &Address) -> Result<Option<EntryMeta>, HolochainError> {
-        let response = self.block_on_ask(Protocol::GetMeta(address.clone()))?;
-        unwrap_to!(response => Protocol::GetMetaResult).clone()
-    }
-
-    fn metas_from_entry(&mut self, entry: &Entry) -> Result<Vec<EntryMeta>, HolochainError> {
-        let response = self.block_on_ask(Protocol::MetasFromEntry(entry.clone()))?;
-        unwrap_to!(response => Protocol::MetasFromEntryResult).clone()
-    }
-
-    fn meta_from_request(
-        &mut self,
-        entry_address: Address,
-        attribute_name: &str,
-    ) -> Result<Option<EntryMeta>, HolochainError> {
-        let response = self.block_on_ask(Protocol::MetaFromRequest {
-            entry_address,
-            attribute_name: attribute_name.to_string(),
-        })?;
-        unwrap_to!(response => Protocol::MetaFromRequestResult).clone()
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct HashTableActor<HT: HashTable> {
-    table: HT,
-}
-
-impl<HT: HashTable> HashTableActor<HT> {
-    /// returns a new HastTablActor struct
-    /// internal use for riker, use new_ref instead
-    fn new(table: HT) -> HashTableActor<HT> {
-        HashTableActor { table }
-    }
-
-    /// actor() for riker
-    fn actor(table: HT) -> BoxActor<Protocol> {
-        Box::new(HashTableActor::new(table))
-    }
-
-    /// props() for riker
-    fn props(table: HT) -> BoxActorProd<Protocol> {
-        Props::new_args(Box::new(HashTableActor::actor), table)
-    }
-
-    /// returns a new actor ref for a new HashTableActor in the main actor system
-    pub fn new_ref(table: HT) -> ActorRef<Protocol> {
-        SYS.actor_of(
-            HashTableActor::props(table),
-            &snowflake::ProcessUniqueId::new().to_string(),
-        ).unwrap()
-    }
-}
-
-impl<HT: HashTable> Actor for HashTableActor<HT> {
-    type Msg = Protocol;
-
-    fn receive(
-        &mut self,
-        context: &Context<Self::Msg>,
-        message: Self::Msg,
-        sender: Option<ActorRef<Self::Msg>>,
-    ) {
-        sender
-            .try_tell(
-                // every Protocol for HashTable maps directly to a method of the same name
-                match message {
-                    Protocol::Setup => Protocol::SetupResult(self.table.setup()),
-
-                    Protocol::Teardown => Protocol::TeardownResult(self.table.teardown()),
-
-                    Protocol::PutEntry(entry) => {
-                        Protocol::PutEntryResult(self.table.put_entry(&entry))
-                    }
-
-                    Protocol::GetEntry(hash) => Protocol::GetEntryResult(self.table.entry(&hash)),
-
-                    Protocol::ModifyEntry { keys, old, new } => {
-                        Protocol::ModifyEntryResult(self.table.modify_entry(&keys, &old, &new))
-                    }
-
-                    Protocol::RetractEntry { keys, entry } => {
-                        Protocol::RetractEntryResult(self.table.retract_entry(&keys, &entry))
-                    }
-
-                    Protocol::AssertMeta(meta) => {
-                        Protocol::AssertMetaResult(self.table.assert_meta(&meta))
-                    }
-
-                    Protocol::GetMeta(key) => Protocol::GetMetaResult(self.table.get_meta(&key)),
-
-                    Protocol::MetasFromEntry(entry) => {
-                        Protocol::MetasFromEntryResult(self.table.metas_from_entry(&entry))
-                    }
-
-                    Protocol::MetaFromRequest {
-                        entry_address,
-                        attribute_name,
-                    } => Protocol::MetaFromRequestResult(
-                        self.table.meta_from_request(entry_address, &attribute_name),
-                    ),
-
-                    _ => unreachable!(),
-                },
-                Some(context.myself()),
-            )
-            .expect("could not tell to HashTableActor sender");
-    }
-}
-
-#[cfg(test)]
-pub mod tests {
-    use super::HashTableActor;
-    use hash_table::{memory::tests::test_table, test_util::standard_suite, HashTable};
-    use holochain_core_types::{
-        actor::Protocol,
-        cas::content::AddressableContent,
-        entry::{test_entry, test_entry_address},
-    };
-    use riker::actors::*;
-    use std::{sync::mpsc, thread};
-
-    /// dummy table actor ref
-    /// every call produces a new actor, not just a new ref to the same actor
-    pub fn test_table_actor() -> ActorRef<Protocol> {
-        HashTableActor::new_ref(test_table())
-    }
-
-    #[test]
-    fn round_trip() {
-        let mut table_actor = test_table_actor();
-
-        assert_eq!(table_actor.entry(&test_entry_address()).unwrap(), None);
-
-        table_actor.put_entry(&test_entry()).unwrap();
-
-        assert_eq!(
-            table_actor.entry(&test_entry().address()).unwrap().unwrap(),
-            test_entry(),
-        );
-    }
-
-    #[test]
-    /// show two things here:
-    /// - we can clone some stateful thing (i.e. actor ref) and mutate one clone and have that
-    ///   consistent across all the clones
-    /// - we can send the cloned stateful thing into threads and have them see a consistent world
-    ///   view without juggling direct message passing through channels
-    fn test_round_trip_threads() {
-        let table_actor = test_table_actor();
-
-        let table_actor_thread = table_actor.clone();
-        let (tx1, rx1) = mpsc::channel();
-        thread::spawn(move || {
-            assert_eq!(
-                table_actor_thread.entry(&test_entry_address()).unwrap(),
-                None
-            );
-            // kick off the next thread
-            tx1.send(true).unwrap();
-        });
-
-        // mutate this clone of the original actor ref
-        let mut table_actor_thread = table_actor.clone();
-        let (tx2, rx2) = mpsc::channel();
-        thread::spawn(move || {
-            rx1.recv().unwrap();
-
-            let entry = test_entry();
-            table_actor_thread.put_entry(&entry).unwrap();
-
-            // push the committed entry through to the next thread
-            tx2.send(entry).unwrap();
-        });
-
-        let table_actor_thread = table_actor.clone();
-        let handle = thread::spawn(move || {
-            let entry = rx2.recv().unwrap();
-            assert_eq!(
-                table_actor_thread.entry(&entry.address()).unwrap().unwrap(),
-                test_entry(),
-            );
-        });
-
-        handle.join().unwrap();
-    }
-
-    #[test]
-    fn test_standard_suite() {
-        standard_suite(&mut test_table_actor());
-    }
-
-}
