+++ core/src/lib.rs
@@ -8,16 +8,14 @@ extern crate rust_base58;
 extern crate serde;
 #[macro_use]
 extern crate serde_json;
-extern crate snowflake;
-#[cfg(test)]
-extern crate test_utils;
-extern crate wasmi;
-#[macro_use]
-extern crate bitflags;
 extern crate futures;
 extern crate riker;
 extern crate riker_default;
 extern crate riker_patterns;
+extern crate snowflake;
+#[cfg(test)]
+extern crate test_utils;
+extern crate wasmi;
 #[macro_use]
 extern crate unwrap_to;
 #[macro_use]
@@ -40,7 +38,6 @@ pub mod action;
 pub mod agent;
 pub mod context;
 pub mod dht;
-pub mod hash_table;
 pub mod instance;
 pub mod logger;
 pub mod nucleus;
