+++ core_types/Cargo.toml
@@ -17,6 +17,7 @@ unwrap_to = "0.1.0"
 lazy_static = "1.1.0"
 riker-patterns = "0.1.7"
 riker-default = "0.1.7"
+bitflags = "1.0"
 
 [dev-dependencies]
 test_utils = { path = "../test_utils"}
