+++ core_types/src/crud_status.rs
@@ -1,4 +1,4 @@
-use holochain_core_types::cas::content::{AddressableContent, Content};
+use cas::content::{AddressableContent, Content};
 
 // @TODO are these the correct key names?
 // @see https://github.com/holochain/holochain-rust/issues/143
@@ -61,15 +61,13 @@ impl AddressableContent for CrudStatus {
 #[cfg(test)]
 mod tests {
     use super::CrudStatus;
-    use holochain_core_types::{
-        cas::{
-            content::{
-                AddressableContent, AddressableContentTestSuite, Content, ExampleAddressableContent,
-            },
-            storage::{test_content_addressable_storage, ExampleContentAddressableStorage},
+    use cas::{
+        content::{
+            AddressableContent, AddressableContentTestSuite, Content, ExampleAddressableContent,
         },
-        eav::eav_round_trip_test_runner,
+        storage::{test_content_addressable_storage, ExampleContentAddressableStorage},
     };
+    use eav::eav_round_trip_test_runner;
 
     #[test]
     /// test the CrudStatus bit flags as ints
